Minor bugfixes and default token stuff
The script will now update the default token for npcs when their token is dropped on the map and updated to match your preferences. This should eliminate the flash of dynamic lighting when a token that hasn't been altered yet is dropped onto the map. Also removed an erroneous log (); left in the code.